refactor, update cra libraries
