+++ src/data/index.js
@@ -1,0 +1,27 @@
+const scale = [
+  'Almost Always',
+  'Very Frequently',
+  'Somewhat Frequently',
+  'Somewhat Infrequently',
+  'Very Infrequently',
+  'Almost Never'
+]
+const questions = [
+  'I could be experiencing some emotion and not be conscious of it until some time later.',
+  'I break or spill things because of carelessness, not paying attention, or thinking of something else.',
+  'I find it difficult to stay focused on what\'s happening in the present.',
+  'I tend to walk quickly to get where I\'m going without paying attention to what I experience along the way.',
+  'I tend not to notice feelings of physical tension or discomfort until they really grab my attention.',
+  'I forget a person\'s name almost as soon as I\'ve been told it for the first time.',
+  'It seems I am "running on automatic," without much awareness of what I\'m doing.',
+  'I rush through activities without being really attentive to them.',
+  'I get so focused on the goal I want to achieve that I lose touch with what I\'m doing right now to get there.',
+  'I do jobs or tasks automatically, without being aware of what I\'m doing.',
+  'I find myself listening to someone with one ear, doing something else at the same time.',
+  'I drive places on "automatic pilot" and then wonder why I went there.',
+  'I find myself preoccupied with the future or the past.',
+  'I find myself doing things without paying attention.',
+  'I snack without being aware that I\'m eating.'
+]
+
+export { scale, questions }
